pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "idealista-scraper"
version = "0.1.0"
description = "Stealthy scraper for Idealista selling apartments in Sevilla"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "parsel>=1.8.0",
    "httpx>=0.25.0",
    "playwright>=1.40.0",
    "playwright-stealth>=1.0.6",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0.0", "pytest-asyncio>=0.21.0"]

[project.scripts]
idealista-scraper = "idealista_scraper.cli:_main"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
